containers/PytorchWildlife/app/src/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.3
fire==0.6.0
fonttools==4.50.0
fsspec==2024.3.1
google-auth==2.29.0
google-auth-oauthlib==1.0.0
gradio==4.8.0
gradio_client==0.7.1
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.10.0
packaging==24.0
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.43
protobuf==3.20.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PytorchWildlife==1.0.2.5
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.36.3
supervision==0.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
tomlkit==0.12.0
toolz==0.12.1
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.66.1
traitlets==5.14.2
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics-yolov5==0.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.1
zipp==3.18.1
